sweet alert added

+++ src/pages/MyServices.jsx
@@ -2,6 +2,7 @@ import React, { useContext, useEffect, useState } from 'react';
 import { AuthContext } from '../Provider/AuthProvider';
 import { Link } from 'react-router';
 import axios from 'axios';
+import Swal from 'sweetalert2';
 
 const MyServices = () => {
     const [myServices, setMyServices] = useState([]);
@@ -15,13 +16,33 @@ const MyServices = () => {
     console.log(myServices);
 
     const handleDelete = (id) => {
-        axios.delete(`http://localhost:3000/delete/${id}`)
-            .then(res => {
-                console.log(res)
-                const filterData = myServices.filter(service => service?._id != id)
-                setMyServices(filterData);
-            })
-            .catch(err => console.log(err))
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                axios.delete(`http://localhost:3000/delete/${id}`)
+                    .then(res => {
+                        console.log(res)
+                        if (res.data.deletedCount) {
+                            const filterData = myServices.filter(service => service?._id != id)
+                            setMyServices(filterData);
+                        }
+                    })
+                    .catch(err => console.log(err))
+                Swal.fire({
+                    title: "Deleted!",
+                    text: "Your file has been deleted.",
+                    icon: "success"
+                });
+            }
+        });
+
     }
     return (
         <div className="overflow-x-auto">
